requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
pydantic_core==2.23.4
firebase-admin==6.6.0
python-dotenv==1.0.1